scraper/requirements.txt
pyiceberg[sql-postgres,s3fs,pyarrow]==0.9.0
psycopg2-binary==2.9.10
